+++ kwimage/structs/boxes.py
@@ -1596,18 +1596,17 @@ class _BoxTransformMixins(object):
 
     def clip(self, x_min, y_min, x_max, y_max, inplace=False):
         """
-        Clip boxes to image boundaries.
-
-        If box is in ltrb format, inplace operation is an option.
+        Clip boxes to boundaries specified as minimum and maximum coordinates.
 
         Args:
-            x_min (int): minimum x-coordinate
-            y_min (int): minimum x-coordinate
-            x_max (int): maximum x-coordinate
-            y_max (int): maximum y-coordinate
+            x_min (Number): minimum x-coordinate
+            y_min (Number): minimum y-coordinate
+            x_max (Number): maximum x-coordinate
+            y_max (Number): maximum y-coordinate
             inplace (bool):
-                if True and possible, perform operation inplace. Defaults to
-                False.
+                if True and possible, perform the operation inplace.
+                This is only available if the data is in ltrb format.
+                Defaults to False.
 
         Returns:
             Boxes: clipped boxes
@@ -1634,27 +1633,12 @@ class _BoxTransformMixins(object):
         if len(new) == 0:
             return new
 
-        if True:
-            impl = self._impl
-            x1, y1, x2, y2 = impl.T(new.data)
-            np.clip(x1, x_min, x_max, out=x1)
-            np.clip(y1, y_min, y_max, out=y1)
-            np.clip(x2, x_min, x_max, out=x2)
-            np.clip(y2, y_min, y_max, out=y2)
-        else:
-            torch = sys.modules.get('torch', None)
-            if torch is not None and torch.is_tensor(new.data):
-                x1, y1, x2, y2 = new.data.t()
-                x1.clamp_(x_min, x_max)
-                y1.clamp_(y_min, y_max)
-                x2.clamp_(x_min, x_max)
-                y2.clamp_(y_min, y_max)
-            else:
-                x1, y1, x2, y2 = new.data.T
-                np.clip(x1, x_min, x_max, out=x1)
-                np.clip(y1, y_min, y_max, out=y1)
-                np.clip(x2, x_min, x_max, out=x2)
-                np.clip(y2, y_min, y_max, out=y2)
+        impl = self._impl
+        x1, y1, x2, y2 = impl.T(new.data)
+        impl.clip(x1, x_min, x_max, out=x1)
+        impl.clip(y1, y_min, y_max, out=y1)
+        impl.clip(x2, x_min, x_max, out=x2)
+        impl.clip(y2, y_min, y_max, out=y2)
         return new
 
     def resize(self, width=None, height=None, inplace=False, about='xy'):
@@ -1770,10 +1754,10 @@ class _BoxTransformMixins(object):
         of each bounding box.
 
         Args:
-            x_left (int | float): xmin pad
-            y_top (int | float): ymin pad
-            x_right (int | float): xmax pad
-            y_bot (int | float): ymax pad
+            x_left (Number): xmin pad
+            y_top (Number): ymin pad
+            x_right (Number): xmax pad
+            y_bot (Number): ymax pad
 
         Returns:
             Boxes : padded boxes
@@ -2940,7 +2924,7 @@ class Boxes(_BoxConversionMixins, _BoxPropertyMixins, _BoxTransformMixins,
         """
         torch = sys.modules.get('torch', None)
         if torch is None:
-            raise Exception('torch is not available')
+            raise Exception('torch has not been imported or is not available')
         data = self.data
         if not torch.is_tensor(data):
             data = torch.from_numpy(data)